fixes for sonarqube code duplication issue
