+++ apps/backend/coverage.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" ?>
-<coverage version="7.14.3" timestamp="1783626700518" lines-valid="3874" lines-covered="3601" line-rate="0.9295" branches-covered="0" branches-valid="0" branch-rate="0" complexity="0">
+<coverage version="7.14.3" timestamp="1783627655376" lines-valid="3875" lines-covered="3602" line-rate="0.9295" branches-covered="0" branches-valid="0" branch-rate="0" complexity="0">
 	<!-- Generated by coverage.py: https://coverage.readthedocs.io/en/7.14.3 -->
 	<!-- Based on https://raw.githubusercontent.com/cobertura/web/master/htdocs/xml/coverage-04.dtd -->
 	<sources>
@@ -83,24 +83,24 @@
 						<line number="14" hits="1"/>
 						<line number="15" hits="1"/>
 						<line number="17" hits="1"/>
-						<line number="19" hits="1"/>
 						<line number="20" hits="1"/>
-						<line number="21" hits="0"/>
-						<line number="23" hits="1"/>
-						<line number="24" hits="1"/>
-						<line number="25" hits="0"/>
-						<line number="26" hits="0"/>
+						<line number="21" hits="1"/>
+						<line number="22" hits="0"/>
+						<line number="25" hits="1"/>
+						<line number="26" hits="1"/>
 						<line number="27" hits="0"/>
-						<line number="30" hits="0"/>
-						<line number="33" hits="1"/>
-						<line number="34" hits="1"/>
-						<line number="36" hits="0"/>
+						<line number="28" hits="0"/>
+						<line number="29" hits="0"/>
+						<line number="32" hits="0"/>
+						<line number="35" hits="1"/>
+						<line number="36" hits="1"/>
 						<line number="38" hits="0"/>
-						<line number="39" hits="0"/>
-						<line number="43" hits="0"/>
-						<line number="47" hits="0"/>
-						<line number="48" hits="0"/>
-						<line number="52" hits="0"/>
+						<line number="40" hits="0"/>
+						<line number="41" hits="0"/>
+						<line number="45" hits="0"/>
+						<line number="49" hits="0"/>
+						<line number="50" hits="0"/>
+						<line number="54" hits="0"/>
 					</lines>
 				</class>
 				<class name="drone.py" filename="apps/backend/app/api/drone.py" complexity="0" line-rate="0.9038" branch-rate="0">
@@ -4132,7 +4132,13 @@
 		</package>
 		<package name="services.tests.database_manager_testing.model_testing" line-rate="0.9952" branch-rate="0" complexity="0">
 			<classes>
-				<class name="conftest.py" filename="services/tests/database_manager_testing/model_testing/conftest.py" complexity="0" line-rate="0.96" branch-rate="0">
+				<class name="conftest.py" filename="services/tests/database_manager_testing/model_testing/conftest.py" complexity="0" line-rate="1" branch-rate="0">
+					<methods/>
+					<lines>
+						<line number="1" hits="1"/>
+					</lines>
+				</class>
+				<class name="db_fixtures.py" filename="services/tests/database_manager_testing/model_testing/db_fixtures.py" complexity="0" line-rate="0.96" branch-rate="0">
 					<methods/>
 					<lines>
 						<line number="1" hits="1"/>
