+++ services/tests/database_manager_testing/model_testing/db_fixtures.py
@@ -1,0 +1,32 @@
+import pytest_asyncio
+from sqlalchemy import event
+from sqlalchemy.ext.asyncio import async_sessionmaker, create_async_engine
+
+from services.database_manager.database import Base
+from services.database_manager.models.drones import Drone
+from services.database_manager.models.flight_summary import FlightSummary
+from services.database_manager.models.telemetry import Telemetry
+from services.database_manager.models.users import User
+
+
+@pytest_asyncio.fixture
+async def engine():
+	engine = create_async_engine('sqlite+aiosqlite:///:memory:')
+
+	@event.listens_for(engine.sync_engine, 'connect')
+	def set_sqlite_pragma(dbapi_connection, connection_record):
+		cursor = dbapi_connection.cursor()
+		cursor.execute('PRAGMA foreign_keys=ON')
+		cursor.close()
+
+	async with engine.begin() as conn:
+		await conn.run_sync(Base.metadata.create_all)
+	yield engine
+	await engine.dispose()
+
+
+@pytest_asyncio.fixture
+async def session(engine):
+	session_maker = async_sessionmaker(engine, expire_on_commit=False)
+	async with session_maker() as s:
+		yield s
